backend/requirements.txt
asgiref==3.7.2
Django==5.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
factory-boy==3.3.0
Faker==20.1.0
freezegun==1.3.1
parameterized==0.9.0
# psycopg==3.1.14
# psycopg-binary==3.1.14
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
setuptools==69.0.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
wheel==0.42.0
